Split Emoji presenter composable from Treehouse presenter setup
This will facilitate running the presenter composable directly in a browser.
