+++ samples/emoji-search/android-composeui/build.gradle
@@ -17,7 +17,7 @@ dependencies {
   implementation libs.google.material
   implementation libs.kotlinx.coroutines.android
   implementation projects.samples.emojiSearch.launcher
-  implementation projects.samples.emojiSearch.presenters
+  implementation projects.samples.emojiSearch.presenterTreehouse
   implementation projects.samples.emojiSearch.schema.widget.protocol
   implementation libs.androidx.activity.compose
   implementation libs.androidx.appCompat
