+++ samples/emoji-search/ios-shared/build.gradle
@@ -15,7 +15,7 @@ kotlin {
     commonMain {
       dependencies {
         implementation projects.samples.emojiSearch.launcher
-        implementation projects.samples.emojiSearch.presenters
+        implementation projects.samples.emojiSearch.presenterTreehouse
         implementation projects.samples.emojiSearch.schema.widget.protocol
         implementation projects.redwoodLayoutUiview
         implementation projects.redwoodTreehouseLazylayoutUiview
