+++ samples/emoji-search/presenter-treehouse/src/commonMain/kotlin/app/cash/zipline/samples/emojisearch/EmojiSearchPresenter.kt
