+++ samples/emoji-search/presenter-treehouse/src/jsMain/kotlin/app/cash/zipline/samples/emojisearch/RealEmojiSearchPresenter.kt
@@ -15,8 +15,13 @@
  */
 package app.cash.zipline.samples.emojisearch
 
+import androidx.compose.runtime.Composable
+import app.cash.redwood.protocol.compose.ProtocolBridge
+import app.cash.redwood.treehouse.TreehouseUi
 import app.cash.redwood.treehouse.ZiplineTreehouseUi
 import app.cash.redwood.treehouse.asZiplineTreehouseUi
+import app.cash.redwood.treehouse.lazylayout.compose.LazyColumn
+import app.cash.redwood.treehouse.lazylayout.compose.items
 import example.schema.compose.EmojiSearchProtocolBridge
 import kotlinx.serialization.json.Json
 
@@ -26,13 +31,41 @@ class RealEmojiSearchPresenter(
 ) : EmojiSearchPresenter {
   override fun launch(): ZiplineTreehouseUi {
     val bridge = EmojiSearchProtocolBridge.create(json)
-    val treehouseUi = EmojiSearchTreehouseUi(
-      hostApi = hostApi,
-      bridge = bridge,
-    )
+    val httpClient = HostHttpClient(hostApi)
+    val lazyColumnProvider = LazyColumnProvider(bridge)
+    val treehouseUi = object : TreehouseUi {
+      @Composable
+      override fun Show() {
+        EmojiSearch(httpClient, lazyColumnProvider)
+      }
+    }
     return treehouseUi.asZiplineTreehouseUi(
       bridge = bridge,
       widgetVersion = 0U,
     )
+  }
+}
+
+private class HostHttpClient(
+  private val hostApi: HostApi,
+) : HttpClient {
+  override suspend fun call(url: String, headers: Map<String, String>): String {
+    return hostApi.httpCall(url, headers)
+  }
+}
+
+private class LazyColumnProvider(
+  private val bridge: ProtocolBridge,
+) : ColumnProvider {
+  @Composable
+  override fun <T> create(
+    items: List<T>,
+    itemContent: @Composable (item: T) -> Unit,
+  ) {
+    bridge.LazyColumn {
+      items(items) { item ->
+        itemContent(item)
+      }
+    }
   }
 }
