+++ settings.gradle
@@ -57,7 +57,8 @@ if (!hasProperty('redwoodNoSamples')) {
   include ':samples:emoji-search:android-views'
   include ':samples:emoji-search:ios-shared'
   include ':samples:emoji-search:launcher'
-  include ':samples:emoji-search:presenters'
+  include ':samples:emoji-search:presenter'
+  include ':samples:emoji-search:presenter-treehouse'
   include ':samples:emoji-search:schema'
   include ':samples:emoji-search:schema:compose'
   include ':samples:emoji-search:schema:compose:protocol'
